+++ ledger-wasm/ledger-v9.template.d.ts
@@ -1,4 +1,4 @@
-<% print(fs.readFileSync('../onchain-runtime-wasm/onchain-runtime-v3.d.ts', 'utf8')); %>
+<% print(fs.readFileSync('../onchain-runtime-wasm/onchain-runtime-v4.d.ts', 'utf8')); %>
 
 /**
  * A zero-knowledge proof.
@@ -92,6 +92,7 @@ export class SignatureEnabled {
   toString(compact?: boolean): string;
   readonly instance: 'signature';
   private type_: 'signature';
+  readonly value: Signature;
 }
 
 export class SignatureErased {
@@ -285,6 +286,15 @@ export type EventDetails =
     vFee: bigint,
     declaredTime: Date,
     blockTime: Date,
+  } | {
+    tag: 'contractLog',
+    address: ContractAddress,
+    entryPoint: Uint8Array | string,
+    loggedItem: {
+      version: number,
+      eventType: LogEventType,
+      data: EncodedStateValue,
+    },
   } |
   // Other variants may be added and some events are not yet supported in this API.
   { tag: string };
@@ -978,9 +988,9 @@ export class Intent<S extends Signaturish, P extends Proofish, B extends Binding
 export class UnshieldedOffer<S extends Signaturish> {
   private constructor();
 
-  static new(inputs: UtxoSpend[], outputs: UtxoOutput[], signatures: Signature[]): UnshieldedOffer<SignatureEnabled>;
+  static new(inputs: UtxoSpend[], outputs: UtxoOutput[], signatures: SignatureEnabled[]): UnshieldedOffer<SignatureEnabled>;
 
-  addSignatures(signatures: Signature[]): UnshieldedOffer<S>;
+  addSignatures(signatures: S[]): UnshieldedOffer<S>;
 
   eraseSignatures(): UnshieldedOffer<SignatureErased>;
 
@@ -988,7 +998,7 @@ export class UnshieldedOffer<S extends Signaturish> {
 
   readonly inputs: UtxoSpend[];
   readonly outputs: UtxoOutput[];
-  readonly signatures: Signature[];
+  readonly signatures: S[];
 }
 
 /**
@@ -1710,8 +1720,11 @@ export class ZswapChainState {
    *
    * Typically, `postBlockUpdate` should be run after any (sequence of)
    * (system)-transaction application(s).
+   *
+   * @param tblock - timestamp of a block last batch of updates was applied at
+   * @param retentionDuration - number of seconds to retain past Merkle tree roots
    */
-  postBlockUpdate(tblock: Date): ZswapChainState;
+  postBlockUpdate(tblock: Date, retentionDuration: bigint): ZswapChainState;
 
   /**
    * Try to apply an {@link ZswapOffer} to the state, returning the updated state