+++ ledger/tests/contract_deployment.rs
@@ -22,6 +22,7 @@ use midnight_ledger::test_utilities::{
     Resolver, test_intents, test_resolver, tx_prove, verifier_key,
 };
 use midnight_ledger::verify::WellFormedStrictness;
+use midnight_ledger_v9 as midnight_ledger;
 use onchain_runtime::state::{ContractOperation, ContractState, StateValue, stval};
 use rand::SeedableRng;
 use rand::rngs::StdRng;
@@ -41,7 +42,7 @@ async fn zero_contract_deploy_balance() {
 
     // Part 1: Deploy
     println!(":: Part 1: Deploy");
-    let count_op = ContractOperation::new(verifier_key(&RESOLVER, "count").await);
+    let count_op = ContractOperation::new(verifier_key(&RESOLVER, "count").await, None);
     let mut contract = ContractState::new(
         stval!([(0u64), (false), (0u64)]),
         HashMap::new().insert(b"count"[..].into(), count_op.clone()),
@@ -84,7 +85,7 @@ async fn non_zero_contract_deploy_balance() {
 
     // Part 1: Deploy
     println!(":: Part 1: Deploy");
-    let count_op = ContractOperation::new(verifier_key(&RESOLVER, "count").await);
+    let count_op = ContractOperation::new(verifier_key(&RESOLVER, "count").await, None);
     let expected_balance = HashMap::new()
         .insert(
             TokenType::Shielded(ShieldedTokenType(HashOutput([0; PERSISTENT_HASH_BYTES]))),